+++ src/compat/hiredis/hiredis_ssl.h
@@ -1,0 +1,5 @@
+#ifndef INTERNAL_LIBHIREDIS
+#include <hiredis/hiredis_ssl.h>
+#else
+#include <hiredis_ssl.h>
+#endif