+++ src/libhiredis-wrapper/async-ctx/async-ctx-creator.cc
@@ -16,30 +16,21 @@
     along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#pragma once
+#include "async-ctx-creator.hh"
 
-#include <chrono>
-#include <filesystem>
-#include <string>
-#include <variant>
-
-#include "flexisip/configmanager.hh"
-
-#include "libhiredis-wrapper/async-ctx/parameters.hh"
-#include "libhiredis-wrapper/redis-auth.hh"
+#include "flexisip/logmanager.hh"
 
 namespace flexisip::redis::async {
 
-struct RedisParameters {
-	std::string domain{};
-	std::variant<redis::auth::None, redis::auth::Legacy, redis::auth::ACL> auth{};
-	int port = 0;
-	std::chrono::seconds mSlaveCheckTimeout{0};
-	bool useSlavesAsBackup = true;
-	std::chrono::seconds mSubSessionKeepAliveTimeout{0};
-	ConnectionParameters connectionParameters{};
+AsyncCtxCreator::AsyncCtxCreator() : mLogPrefix(LogManager::makeLogPrefixForInstance(this, "AsyncCtxCreator")) {
+}
 
-	static RedisParameters fromRegistrarConf(GenericStruct const*);
-};
+AsyncContextPtr AsyncCtxCreator::createAsyncCtx(const std::string_view& address, int port) {
+	AsyncContextPtr ctx{redisAsyncConnect(address.data(), port)};
+	if (ctx == nullptr) {
+		throw std::bad_alloc{};
+	}
 
+	return ctx;
+}
 } // namespace flexisip::redis::async